Replace "Unexpected end of NAME" with read property
Existing code was throwing an error, even though it read the ICS correctly.  When it hit the EOF, it didn't return the calendar property that it had finally read in, it threw an error.  Now it returns whatever data was read in and it allows the other logic in the code to determine if it is an expected value or not.